Separation
He abandoned the "Hand Detection" method because it does not separate the right and left hand. MEDIAPIPE has a "Holistic" method, thanks to which you can take the left and right hand separately, as well as face tracking, but a small FPS. So far, this is the only method I need to implement sign language.